fix: move buildpacker imager code into own go file
