+++ internal/pipe/docker/api_docker.go
@@ -3,7 +3,6 @@ package docker
 import (
 	"context"
 	"fmt"
-	"strings"
 )
 
 func init() {
@@ -66,44 +65,6 @@ func (i dockerImager) buildCommand(images, flags []string) []string {
 	for _, image := range images {
 		base = append(base, "-t", image)
 	}
-	base = append(base, flags...)
-	return base
-}
-
-type buildPackImager struct {
-}
-
-func (i buildPackImager) Push(ctx context.Context, image string, flags []string) error {
-	if err := runCommand(ctx, ".", "docker", "push", image); err != nil {
-		return fmt.Errorf("failed to push %s: %w", image, err)
-	}
-	return nil
-}
-
-func (i buildPackImager) Build(ctx context.Context, root string, images, flags []string) error {
-	if err := runCommand(ctx, root, "pack", i.buildCommand(images, flags)...); err != nil {
-		return fmt.Errorf("failed to build %s: %w", images[0], err)
-	}
-	return nil
-}
-
-func (i buildPackImager) buildCommand(images, flags []string) []string {
-	base := []string{"build", images[0]}
-	for j := 1; j < len(images); j++ {
-		base = append(base, "-t", images[j])
-	}
-
-	builderConfigured := false
-	for _, flag := range flags {
-		if strings.HasPrefix(flag, "-B") || strings.HasPrefix(flag, "--builder") {
-			builderConfigured = true
-		}
-	}
-
-	if !builderConfigured {
-		flags = append(flags, "--builder=gcr.io/buildpacks/builder:v1")
-	}
-
 	base = append(base, flags...)
 	return base
 }
